backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

asyncpg==0.29.0

python-dotenv==1.0.1

httpx==0.27.0
yclients-api==0.0.7
pandas==2.2.3
ujson==5.10.0
